Add API to preload specific unmanaged DLLs
Unmanaged DLLs that are loaded outide the assembly load context are not
found when hot reloading (load in memory) is enabled. This happens when
NativeLibrary.Load(String, Assembly, DllImportSearchPath?) is used.

This change adds an API that lets consumers of this package configure a
set of unmanaged libraries to preload before any other resolving
happens.

Fixes #148